+++ pgmacs.el
@@ -2,8 +2,8 @@
 
 ;; Copyright (C) 2023-2025 Eric Marsden
 ;; Author: Eric Marsden <eric.marsden@risk-engineering.org>
-;; Version: 0.23
-;; Package-Requires: ((emacs "29.1") (pg "0.51"))
+;; Version: 0.24
+;; Package-Requires: ((emacs "29.1") (pg "0.53"))
 ;; URL: https://github.com/emarsden/pgmacs/
 ;; Keywords: data, PostgreSQL, database
 ;; SPDX-License-Identifier: GPL-3.0-or-later